Funct

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,5 @@
 #include "function_pointers.h"
+#include <stddef.h>
 /**
  * array_iterator - Function that iterates an array
  * @array: Character that stores array value
@@ -7,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array && action)
 	{
